Use schemas to correctly build messages
According to the specification for [UPnP Devices][1], the arguments for
the control actions need to match the order of arguments as desribed in
the schema. To achieve that, we retrieve the schemas when searching for
a compitable gateway and use the appropriate schema for each action to
build the arguments properly.

[1]: https://openconnectivity.org/upnp-specs/UPnP-arch-DeviceArchitecture-v2.0-20200417.pdf