Warn about extraneous arguments, do not panic
To panic is not a nice way of handling unintended behavior of a library.
In case the router presents us with arguments we do not know about,
issue a warning via the logger, but do not bail out. Proceed as if
everything is okay and let the router decide whether it was fatal.